refactor: move ff-admin creation to forge
